services/notification-service/requirements.txt
aio_pika==9.4.0
aiogram
pydantic==2.5.0
python-dotenv==1.0.0